refactor: extract shared conversion helpers to reduce duplication
Move StringSetToSlice, StringMapToAnyMap, and AnyMapToStringMap from
accesskey and tenant packages into a shared internal/models/convert
package. Extract configureSettings helper in tenant resource to
eliminate duplicated settings configuration block.
